requirements.txt
beautifulsoup4==4.14.3
faiss-cpu==1.13.2
fastapi==0.128.0
groq==1.0.0
mlflow==3.8.1
numpy==2.4.1
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
sentence-transformers==5.2.0
tqdm==4.67.1
uvicorn==0.40.0